add local changed, rename branch
