+++ docker/pathlinker/outk-100-ranked-edges.txt
@@ -1,0 +1,18 @@
+#tail	head	KSP index	path cost
+S2	T3	1	5.00000e-01
+A	E	2	1.25000e-01
+S1	A	2	1.25000e-01
+E	T1	2	1.25000e-01
+S2	B	3	1.25000e-01
+F	T2	3	1.25000e-01
+B	F	3	1.25000e-01
+S2	C	4	1.25000e-01
+C	F	4	1.25000e-01
+C	G	5	1.25000e-01
+G	T2	5	1.25000e-01
+G	T3	6	1.25000e-01
+E	F	7	6.25000e-02
+B	C	8	6.25000e-02
+F	E	11	6.25000e-02
+G	F	13	6.25000e-02
+F	A	16	3.12500e-02
